refactor csv
